Optimizing config by removing pipeline Ids
