+++ azure-pipelines.yml
@@ -28,8 +28,8 @@ steps:
   displayName: "Package CLI"
 
 - script: |
-    echo "./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --docker-pipeline-id $DOCKER_PIPELINE_ID --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --hld-pipeline-id $HLD_PIPELINE_ID --src-pipeline-id $SRC_PIPELINE_ID --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME"
-    ./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --docker-pipeline-id $DOCKER_PIPELINE_ID --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --hld-pipeline-id $HLD_PIPELINE_ID --src-pipeline-id $SRC_PIPELINE_ID --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME
+    echo "./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME"
+    ./cli-linux init --azure-org $AZURE_ORG --azure-project $AZURE_PROJECT --manifest $GITHUB_MANIFEST --github-manifest-username $GITHUB_MANIFEST_USERNAME --storage-account-key $(STORAGE_ACCOUNT_KEY) --storage-account-name $STORAGE_ACCOUNT_NAME --storage-partition-key $STORAGE_PARTITION_KEY --storage-table-name $STORAGE_TABLE_NAME
   displayName: "Set Config"
   
 - script: |
