+++ src/config.ts
@@ -2,12 +2,9 @@ export let config = {
   AZURE_ORG: "",
   AZURE_PIPELINE_ACCESS_TOKEN: "",
   AZURE_PROJECT: "",
-  DOCKER_PIPELINE_ID: 0,
   GITHUB_MANIFEST_USERNAME: "",
-  HLD_PIPELINE_ID: 0,
   MANIFEST: "",
   MANIFEST_ACCESS_TOKEN: "",
-  SRC_PIPELINE_ID: 0,
   STORAGE_ACCOUNT_KEY: "",
   STORAGE_ACCOUNT_NAME: "",
   STORAGE_PARTITION_KEY: "",
